+++ src/game/ExitPoint.cpp
@@ -18,7 +18,7 @@
 #include "ExitPoint.h"
 #include <iostream>
 
-ExitPoint::ExitPoint(float x, float y, float w, float h)
+ExitPoint::ExitPoint(double x, double y, double w, double h)
 :Sensor(x, y, w, h)
 {
 }
